@@ -4,37 +4,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.3.0",
-    "@emotion/babel-preset-css-prop": "^11.2.0",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "axios": "^0.21.4",
-    "craco-less": "^1.20.0",
-    "i18next": "^21.2.0",
-    "i18next-browser-languagedetector": "^6.1.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "axios": "1.6.0",
+    "i18next": "^23.8.2",
+    "i18next-browser-languagedetector": "^7.2.0",
     "ics": "^2.31.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-i18next": "^11.12.0",
-    "react-markdown": "^7.0.1",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "react-router-hash-link": "^2.4.3",
-    "react-scripts": "4.0.3",
-    "remark-gfm": "^2.0.0",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^14.0.1",
+    "react-markdown": "^9.0.1",
+    "react-router": "^6.22.0",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "^5.0.1",
+    "remark-gfm": "^4.0.0",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "BROWSER=none PORT=3001 craco start",
@@ -61,8 +46,19 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@craco/craco": "^7.1.0",
+    "@emotion/babel-preset-css-prop": "^11.11.0",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@types/node": "^20.11.16",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^18.2.18",
     "@types/react-helmet": "^6.1.1",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/react-router-hash-link": "^2.4.2"
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-router-hash-link": "^2.4.2",
+    "craco-less": "^3.0.1",
+    "typescript": "^5",
+    "webpack": "^5.0.0"
   }
 }
