@@ -1,12 +1,10 @@
-import React from 'react'
-import { SectionT } from './Section'
-import { Link, useHistory, useLocation } from 'react-router-dom'
-import { ChapterT } from './Chapter'
-import { HashLink } from 'react-router-hash-link'
-import { useState } from 'react'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { StartPage } from '../pages/HomePage '
-import { CalendarPageT } from '../pages/CalendarPage'
+import React, {useState} from 'react'
+import {SectionT} from './Section'
+import {Link, useLocation, useNavigate} from 'react-router-dom'
+import {ChapterT} from './Chapter'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
+import {StartPage} from '../pages/HomePage '
+import {CalendarPageT} from '../pages/CalendarPage'
 
 type Props = {
     sections: Array<SectionT>
@@ -18,7 +16,7 @@ function Navigation(props: Props) {
 
     const [navbarOpen, setNavbarOpen] = useState(false)
     const location = useLocation()
-    const history = useHistory()
+    const navigate = useNavigate()
 
     const sections = props.sections
     const [checkedState, setCheckedState] = useState(
@@ -30,7 +28,7 @@ function Navigation(props: Props) {
             index === sectionNav ? !item : false
         );
         setCheckedState(updatedCheckedState)
-        history.push('/' + section.slug)
+        navigate('/' + section.slug)
     }
 
     const handleToggle = () => {
@@ -40,54 +38,57 @@ function Navigation(props: Props) {
     function chapterList(section: SectionT) {
         const chapters = section.chapters
         const chapterItems = chapters.map(function (chapter: ChapterT) {
-            var isActive = location.hash.replace('#', '') === chapter.slug
-            var className = isActive ? 'active' : ''
+            const isActive = location.hash.replace('#', '') === chapter.slug
+            const className = isActive ? 'active' : ''
             return <li key={chapter.slug_with_section} className="subMenu" onClick={handleToggle}>
-                <HashLink to={chapter.slug_with_section} className={className}>{chapter.menu_name}</HashLink>
+                <Link to={chapter.slug_with_section} className={className}>{chapter.menu_name}</Link>
             </li>
         })
+
         return <ul className="accordion_sub-menu">
             {chapterItems}
         </ul>
     }
 
     const sectionList = sections.map(function (section: SectionT, index: number) {
-        var isActive = location.pathname.replace('/', '') === section.slug
-        var className = isActive ? 'active' : ''
-        return <>
-            <li key={section.slug} className={className}>
-                <input
-                    type="checkbox"
-                    name="tabs"
-                    id={section.slug}
-                    className={`accordion_input ${className}`}
-                    checked={checkedState[index]}
-                    onChange={() => handleOnChange(index, section)}
-                />
-                <label htmlFor={section.slug} className={`accordion_label ${className}`}>
-                    {section.menu_name}
-                </label>
-                {chapterList(section)}
-            </li>
-        </>
+        const isActive = location.pathname.replace('/', '') === section.slug
+        const className = isActive ? 'active' : ''
+        return <li key={section.slug} className={className}>
+            <input
+                type="checkbox"
+                name="tabs"
+                id={section.slug}
+                className={`accordion_input ${className}`}
+                checked={checkedState[index]}
+                onChange={() => handleOnChange(index, section)}
+            />
+            <label htmlFor={section.slug} className={`accordion_label ${className}`}>
+                {section.menu_name}
+            </label>
+            {chapterList(section)}
+        </li>
     })
+
     const startPage = props.startPage
     const calendarPage = props.calendarPage
     const isCalendar = location.pathname === '/calendar'
-    var calendarActive = isCalendar ? 'active' : ''
+    const calendarActive = isCalendar ? 'active' : ''
     const isHome = location.pathname === '/'
-    var homeActive = isHome ? 'active' : ''
+    const homeActive = isHome ? 'active' : ''
+
     return <nav className="header-nav">
         <div className="toggle-btn">
-            <i onClick={handleToggle}><FontAwesomeIcon icon="bars" /></i>
+            <i onClick={handleToggle}><FontAwesomeIcon icon="bars"/></i>
         </div>
         <div className={`header-nav-content ${navbarOpen ? "showMenu" : ""}`}>
             <ul className={`menuItems ${navbarOpen ? "showMenu" : ""}`}>
                 <li key="home">
-                    <Link to="/" className={homeActive} onClick={() => setNavbarOpen(!navbarOpen)}>{startPage.menu_name}</Link>
+                    <Link to="/" className={homeActive}
+                          onClick={() => setNavbarOpen(!navbarOpen)}>{startPage.menu_name}</Link>
                 </li>
                 <li key="calendar">
-                    <Link to="/calendar" className={calendarActive} onClick={() => setNavbarOpen(!navbarOpen)}>{calendarPage.menu_name}</Link>
+                    <Link to="/calendar" className={calendarActive}
+                          onClick={() => setNavbarOpen(!navbarOpen)}>{calendarPage.menu_name}</Link>
                 </li>
                 {sectionList}
             </ul>
